Split out model/direct.go
